--- code/dt_archapi_utils/__init__.py
@@ -1,1 +1,0 @@
-from .arch_client import ArchAPIClient
