--- code/dt_class_utils/app_status.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-
-class AppStatus(Enum):
-    INITIALIZING = 10
-    RUNNING = 20
-    TERMINATING = 30
-    KILLING = 40
-    DONE = 100
